Modules for writing Excel sheets and for managing HWItems through a consistent interface
I had the Excel sheet writer mostly working in a "scratch" folder, but I finally put it in the library for general use. I added a bit of code to use the ImageFont module out of PIL (Python Image Library) to get a better estimate of the width of columns in Excel sheets, instead of merely counting characters. It appears to be working really well; the sheets come out looking reallly nice.

The new HWItem module is there to consolidate a lot of various bits of code that do things with Items but not in a really consistent manner. It serves as a go-between between the format things come out of the HWDB REST API as, and the format the Encoder puts things in after it reads a sheet.