Fixing Urbas's merge
I'm still not sure this will be right.

I think what happened was that he "merged" in such a way that he kept his versions of files instead of taking mine, and now it never seems to see my branch and his branch as different, even though THEY ARE!! I manually deleted his /lib and copied mine, but it doesn't show up in changes. So I don't know what is going to happen now.